requirements.txt
gunicorn==20.1.0
pandas==1.3.3
Pillow==8.3.2
streamlit==0.88.0
nltk == 3.5
textblob ==0.15.3
vaderSentiment == 3.3.2
plotly==5.3.1
<<<<<<< HEAD
emoji_translate


=======
git+git://github.com/twintproject/twint.git@origin/master#egg=twint
nest_asyncio
requests==2.26.0
joblib==0.16.0
emoji-translate==0.1.0
wordcloud==1.8.1
>>>>>>> a455a92746df965191edda8898a3d53ab106c1ef
